+++ KTW_DX/GameEngineContents/Player.cpp
@@ -14,7 +14,7 @@ Player::~Player()
 
 void Player::Start()
 {
-	MainSpriteRenderer = CreateComponent<GameEngineSpriteRenderer>(static_cast<int>(RenderOrder::Play));
+	MainSpriteRenderer = CreateComponent<GameEngineSpriteRenderer>(RenderOrder::Play);
 	MainSpriteRenderer->CreateAnimation("CupHead_Idle", "Idle");
 	MainSpriteRenderer->ChangeAnimation("CupHead_Idle");
 	MainSpriteRenderer->AutoSpriteSizeOn();
